+++ pom.xml
@@ -35,7 +35,8 @@
 		<module>spring-cloud-gcp-pubsub-stream-binder</module>
 		<module>spring-cloud-gcp-data-spanner</module>
 		<module>spring-cloud-gcp-data-datastore</module>
-	</modules>
+        <module>spring-cloud-gcp-security-iap</module>
+    </modules>
 
 	<properties>
 		<maven.compiler.source>1.8</maven.compiler.source>
