+++ spring-cloud-gcp-autoconfigure/pom.xml
@@ -193,6 +193,13 @@
             <optional>true</optional>
         </dependency>
 
+        <!-- Security -->
+        <dependency>
+            <groupId>org.springframework.cloud</groupId>
+            <artifactId>spring-cloud-gcp-security-iap</artifactId>
+            <version>1.1.0.BUILD-SNAPSHOT</version>
+        </dependency>
+
         <!-- Test -->
         <dependency>
             <groupId>commons-io</groupId>
