+++ spring-cloud-gcp-autoconfigure/src/main/resources/META-INF/additional-spring-configuration-metadata.json
@@ -28,6 +28,11 @@
       "name": "spring.cloud.gcp.storage.enabled",
       "type": "java.lang.Boolean",
       "description": "Auto-configure Google Cloud Storage components.",
+    },
+    {
+      "name": "spring.cloud.gcp.security.iap.enabled",
+      "type": "java.lang.Boolean",
+      "description": "Auto-configure BLAH",
       "defaultValue": true
     }
   ]
