+++ spring-cloud-gcp-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -9,7 +9,8 @@ org.springframework.cloud.gcp.autoconfigure.sql.GcpCloudSqlAutoConfiguration,\
 org.springframework.cloud.gcp.autoconfigure.storage.GcpStorageAutoConfiguration,\
 org.springframework.cloud.gcp.autoconfigure.trace.StackdriverTraceAutoConfiguration,\
 org.springframework.cloud.gcp.autoconfigure.datastore.DatastoreRepositoriesAutoConfiguration,\
-org.springframework.cloud.gcp.autoconfigure.spanner.SpannerRepositoriesAutoConfiguration
+org.springframework.cloud.gcp.autoconfigure.spanner.SpannerRepositoriesAutoConfiguration,\
+org.springframework.cloud.gcp.autoconfigure.security.IapAuthenticationAutoConfiguration
 
 org.springframework.cloud.bootstrap.BootstrapConfiguration=\
 org.springframework.cloud.gcp.autoconfigure.config.GcpConfigBootstrapConfiguration
