+++ spring-cloud-gcp-samples/spring-cloud-gcp-security-iap-sample/src/main/resources/application.properties
@@ -1,0 +1,1 @@
+spring.cloud.gcp.security.iap.enabled=true
