+++ spring-cloud-gcp-samples/spring-cloud-gcp-security-iap-sample/src/main/resources/logback-spring.xml
@@ -3,7 +3,7 @@
     <include resource="org/springframework/boot/logging/logback/defaults.xml"/>
     <include resource="org/springframework/boot/logging/logback/console-appender.xml" />
 
-    <root level="INFO">
+    <root level="TRACE">
         <!-- If running in GCP, remove the CONSOLE appender otherwise logs will be duplicated. -->
         <appender-ref ref="CONSOLE"/>
     </root>
